requirements.txt
torch==1.9.1
trimesh==3.9.33
tqdm==4.62.3
scipy==1.7.1
scikit-learn==1.0
scikit-image==0.18.3
pyyaml==6.0
pillow==8.3.1
pandas==1.3.4
plyfile==0.7.4
h5py==3.4.0
imageio==2.9.0
open3d
matplotlib==3.3.4